requirements.txt
flask
flask-sqlalchemy
sqlalchemy
flask-cors
PyJWT
PyMySQL
passlib
lepl
flask-restplus